Move LightChain-related tests out of test_peer.py
Also refactor some helper methods and add docstrings to make the tests
easier to follow.
